backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
crispy-bootstrap5==0.7
cycler==0.12.1
diff-match-patch==20241021
distlib==0.3.9
dj-database-url==3.0.1
Django==4.2.1
django-admin-multiple-choice-list-filter==0.1.1
django-cors-headers==4.7.0
django-crispy-forms==2.0
django-import-export==4.4.0
djangorestframework==3.16.0
filelock==3.16.1
fonttools==4.54.1
google-api-core==2.20.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gspread==6.1.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.9.2
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
Pillow==10.0.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.50
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
seaborn==0.13.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
tablib==3.9.0
text-unidecode==1.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
